Ako/ fix bot package circular dependecies (#10687)
* chore: move types to prevent circular dependency

* fix: temporarily use any to prevent circular deps
